requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
elevenlabs>=0.2.26
SpeechRecognition>=3.8.1
pyvts>=0.0.4
requests>=2.26.0
psutil>=5.8.0
pyinstaller>=5.0.0
vosk>=0.3.45
PyAudio>=0.2.13
numpy>=1.24.0 